Chore: remove relative path build flags

+++ internal/RVM/core/system/game/graphic/image/set.go
@@ -1,8 +1,7 @@
 package image
 
 /*
-#cgo CFLAGS: -I./../../../../sdl/include
-#cgo LDFLAGS: -L./../../../../sdl/lib -lSDL2 -lSDL2main -lSDL2_image
+#cgo LDFLAGS: -lSDL2 -lSDL2main -lSDL2_image
 
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_image.h>
@@ -12,7 +11,7 @@ import (
 	"log"
 )
 
-func (i *Image) SetImageAlpha(name string, alpha int) {
+func (i *Image) SetImageAlpha(name string, alpha uint8) {
 	i.lock.Lock()
 	defer i.lock.Unlock()
 	if image, ok := i.images[name]; !ok {
